requirements.txt
openai~=1.59.7
Flask~=3.1.0
python-dotenv~=1.0.1
fastapi~=0.115.6
pydantic~=2.10.5
uvicorn~=0.34.0
langchain~=0.3.14
transformers~=4.48.0